Split automation domain record models
