+++ desktop/src/lib/domain/automations/local-executor.ts
@@ -2,7 +2,7 @@ import type {
   AutomationRepoRootRecord,
   AutomationRunClaimRecord,
   AutomationWorkspaceRecord,
-} from "./records";
+} from "./local-executor-records";
 
 export interface LocalAutomationRepositoryIdentity {
   provider: string;
