+++ desktop/src/lib/domain/automations/target-records.ts
@@ -1,0 +1,13 @@
+export interface AutomationTargetRepoConfigRecord {
+  gitOwner: string;
+  gitRepoName: string;
+  configured: boolean;
+}
+
+export interface AutomationTargetCloudWorkspaceRecord {
+  repo: {
+    provider: string;
+    owner: string;
+    name: string;
+  };
+}
